// src/views/detail/album-detail/store/type.ts
export type albumDetailType = {
  album: {
    songs: Array<any>
    paid: boolean
    onSale: boolean
    mark: number
    awardTags: any
    artists: Array<{
      img1v1Id: number
      topicPerson: number
      picId: number
      musicSize: number
      albumSize: number
      briefDesc: string
      picUrl: string
      img1v1Url: string
      followed: boolean
      trans: string
      alias: Array<any>
      name: string
      id: number
      img1v1Id_str: string
    }>
    copyrightId: number
    picId: number
    artist: {
      img1v1Id: number
      topicPerson: number
      picId: number
      musicSize: number
      albumSize: number
      briefDesc: string
      picUrl: string
      img1v1Url: string
      followed: boolean
      trans: string
      alias: Array<string>
      name: string
      id: number
      picId_str: string
      img1v1Id_str: string
    }
    briefDesc: string
    publishTime: number
    company: string
    picUrl: string
    commentThreadId: string
    blurPicUrl: string
    companyId: number
    pic: number
    status: number
    subType: string
    alias: Array<any>
    description: string
    tags: string
    name: string
    id: number
    type: string
    size: number
    picId_str: string
    info: {
      commentThread: {
        id: string
        resourceInfo: {
          id: number
          userId: number
          name: string
          imgUrl: string
          creator: any
          encodedId: any
          subTitle: any
          webUrl: any
        }
        resourceType: number
        commentCount: number
        likedCount: number
        shareCount: number
        hotCount: number
        latestLikedUsers: any
        resourceId: number
        resourceOwnerId: number
        resourceTitle: string
      }
      latestLikedUsers: any
      liked: boolean
      comments: any
      resourceType: number
      resourceId: number
      commentCount: number
      likedCount: number
      shareCount: number
      threadId: string
    }
  }
  songs: Array<{
    rtUrls: Array<any>
    ar: Array<{
      id: number
      name: string
      alia: Array<string>
    }>
    al: {
      id: number
      name: string
      picUrl: string
      pic_str: string
      pic: number
    }
    st: number
    noCopyrightRcmd: any
    songJumpInfo: any
    no: number
    fee: number
    djId: number
    mv: number
    t: number
    v: number
    cd: string
    rtype: number
    rurl: any
    pst: number
    alia: Array<any>
    pop: number
    rt: string
    mst: number
    cp: number
    crbt: any
    cf: string
    dt: number
    h: any
    sq: any
    hr: any
    l: any
    rtUrl: any
    ftype: number
    a: any
    m: any
    name: string
    id: number
    privilege: {
      id: number
      fee: number
      payed: number
      st: number
      pl: number
      dl: number
      sp: number
      cp: number
      subp: number
      cs: boolean
      maxbr: number
      fl: number
      toast: boolean
      flag: number
      preSell: boolean
      playMaxbr: number
      downloadMaxbr: number
      maxBrLevel: string
      playMaxBrLevel: string
      downloadMaxBrLevel: string
      plLevel: string
      dlLevel: string
      flLevel: string
      rscl: any
      freeTrialPrivilege: {
        resConsumable: boolean
        userConsumable: boolean
        listenType: any
        cannotListenReason: any
        playReason: any
      }
      rightSource: number
      chargeInfoList: Array<{
        rate: number
        chargeUrl: any
        chargeMessage: any
        chargeType: number
      }>
    }
  }>
}
export type IAlbumDetailState = {
  albumDetail: albumDetailType
}